+++ package/src/index.ts
@@ -5,4 +5,8 @@ export { exitPreview } from "./exitPreview";
 export { PrismicPreview } from "./PrismicPreview";
 export { enableAutoPreviews } from "./enableAutoPreviews";
 export { redirectToPreviewURL } from "./redirectToPreviewURL";
-export type { NextContextLike, CreateClientConfig } from "./types";
+export type {
+	NextContextLike,
+	CreateClientConfig,
+	PreviewConfig,
+} from "./types";
