+++ package/src/types.ts
@@ -1,4 +1,6 @@
-import { PreviewData, NextApiRequest } from "next";
+import { PreviewData, NextApiRequest, NextApiResponse } from "next";
+import { LinkResolverFunction } from "@prismicio/helpers";
+import { Client } from "@prismicio/client";
 
 export type NextContextLike<TPreviewData extends PreviewData = PreviewData> = {
 	previewData?: TPreviewData;
@@ -7,4 +9,16 @@ export type NextContextLike<TPreviewData extends PreviewData = PreviewData> = {
 export type CreateClientConfig = {
 	context?: NextContextLike;
 	req?: NextApiRequest;
+};
+
+export type PreviewConfig = {
+	req: {
+		query: NextApiRequest["query"];
+	};
+	res: {
+		redirect: NextApiResponse["redirect"];
+	};
+	client: Client;
+	linkResolver?: LinkResolverFunction;
+	defaultURL?: string;
 };
